+++ src/main/java/wooteco/subway/dao/LineDao.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain.line;
+package wooteco.subway.dao;
 
 import java.util.HashMap;
 import java.util.List;
@@ -8,6 +8,7 @@ import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
+import wooteco.subway.domain.Line;
 import wooteco.subway.web.exception.SubwayHttpException;
 
 @Repository
