+++ src/main/java/wooteco/subway/dao/SectionDao.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain.section;
+package wooteco.subway.dao;
 
 import java.util.List;
 import java.util.Optional;
@@ -7,6 +7,7 @@ import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
+import wooteco.subway.domain.Section;
 import wooteco.subway.web.exception.SubwayHttpException;
 
 @Repository
