+++ src/main/java/wooteco/subway/dao/StationDao.java
@@ -1,16 +1,14 @@
-package wooteco.subway.domain.station;
+package wooteco.subway.dao;
 
 import java.util.Collection;
-import java.util.Collections;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 import org.springframework.dao.DuplicateKeyException;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
+import wooteco.subway.domain.Station;
 import wooteco.subway.web.exception.SubwayHttpException;
 
 @Repository
