+++ src/main/java/wooteco/subway/domain/Line.java
@@ -1,19 +1,21 @@
-package wooteco.subway.domain.station;
+package wooteco.subway.domain;
 
 import java.util.Objects;
 
-public class Station {
+public class Line {
 
     private final Long id;
     private final String name;
+    private final String color;
 
-    public Station(String name) {
-        this(null, name);
+    public Line(String name, String color) {
+        this(null, name, color);
     }
 
-    public Station(Long id, String name) {
+    public Line(Long id, String name, String color) {
         this.id = id;
         this.name = name;
+        this.color = color;
     }
 
     public Long getId() {
@@ -24,6 +26,10 @@ public class Station {
         return name;
     }
 
+    public String getColor() {
+        return color;
+    }
+
     @Override
     public boolean equals(Object o) {
         if (this == o) {
@@ -32,13 +38,13 @@ public class Station {
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        Station station = (Station) o;
-        return Objects.equals(id, station.id) && Objects.equals(name, station.name);
+        Line line = (Line) o;
+        return Objects.equals(id, line.id) && Objects.equals(name, line.name)
+                && Objects.equals(color, line.color);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(id, name);
+        return Objects.hash(id, name, color);
     }
 }
-
