+++ src/main/java/wooteco/subway/domain/Station.java
@@ -1,21 +1,19 @@
-package wooteco.subway.domain.line;
+package wooteco.subway.domain;
 
 import java.util.Objects;
 
-public class Line {
+public class Station {
 
     private final Long id;
     private final String name;
-    private final String color;
 
-    public Line(String name, String color) {
-        this(null, name, color);
+    public Station(String name) {
+        this(null, name);
     }
 
-    public Line(Long id, String name, String color) {
+    public Station(Long id, String name) {
         this.id = id;
         this.name = name;
-        this.color = color;
     }
 
     public Long getId() {
@@ -26,10 +24,6 @@ public class Line {
         return name;
     }
 
-    public String getColor() {
-        return color;
-    }
-
     @Override
     public boolean equals(Object o) {
         if (this == o) {
@@ -38,13 +32,13 @@ public class Line {
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        Line line = (Line) o;
-        return Objects.equals(id, line.id) && Objects.equals(name, line.name)
-                && Objects.equals(color, line.color);
+        Station station = (Station) o;
+        return Objects.equals(id, station.id) && Objects.equals(name, station.name);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(id, name, color);
+        return Objects.hash(id, name);
     }
 }
+
