+++ src/main/java/wooteco/subway/service/LineService.java
@@ -3,8 +3,8 @@ package wooteco.subway.service;
 import java.util.List;
 import org.springframework.dao.IncorrectResultSizeDataAccessException;
 import org.springframework.stereotype.Service;
-import wooteco.subway.domain.line.Line;
-import wooteco.subway.domain.line.LineDao;
+import wooteco.subway.domain.Line;
+import wooteco.subway.dao.LineDao;
 import wooteco.subway.web.exception.NotFoundException;
 
 @Service
