+++ src/main/java/wooteco/subway/web/controller/LineController.java
@@ -14,7 +14,7 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import wooteco.subway.service.LineFacade;
+import wooteco.subway.facade.LineFacade;
 import wooteco.subway.web.dto.LineRequest;
 import wooteco.subway.web.dto.LineResponse;
 import wooteco.subway.web.dto.LineUpdateRequest;
