+++ src/main/java/wooteco/subway/web/controller/SectionController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import wooteco.subway.service.SectionFacade;
+import wooteco.subway.facade.SectionFacade;
 import wooteco.subway.web.dto.SectionRequest;
 
 @RestController
