+++ src/main/java/wooteco/subway/web/controller/StationController.java
@@ -13,8 +13,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import wooteco.subway.domain.station.Station;
-import wooteco.subway.service.StationFacade;
+import wooteco.subway.domain.Station;
+import wooteco.subway.facade.StationFacade;
 import wooteco.subway.web.dto.StationRequest;
 import wooteco.subway.web.dto.StationResponse;
