+++ src/main/java/wooteco/subway/web/dto/LineRequest.java
@@ -3,7 +3,7 @@ package wooteco.subway.web.dto;
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
-import wooteco.subway.domain.line.Line;
+import wooteco.subway.domain.Line;
 
 public class LineRequest {
